Fix Solaris compatibility
Add a workaround for the lack of -mmin in Solarises find, check for
gfind from opencsw and failing that use perl, which Solaris ships with
and is widely available on many other systems.

If neither a working find nor perl is available, print a warning and
turn off remote status checking.

Use eval in async_run to run the command, the old bash in Solaris 5.10
does not recognize the previous syntax.

Use command -v instead of which to check for the existance of rbenv in
the Crunch and Solarized_Extravagant themes. command -v is the POSIX-ly
correct way to check for the existance of a command, and which does not
set the exit code on Solaris.