Fix ignored GIT_PROMPT_WITH_VIRTUAL_ENV
If Anaconda is installed the conda environment is shown in the prompt
even though GIT_PROMPT_WITH_VIRTUAL_ENV=0 is set. The env variable was
used only in git_prompt_config() function, not updatePrompt() function.
The env variable check was added to gp_add_virtualenv_to_prompt() which
is called by both functions.