Use raw ASCII escapes for colors
If you output colors in a shell function, then the bash nonprinting
character escapes \[ and \] don't work -- they get output as literals.
Instead, you have to use raw ASCII escapes \001 and \002, which work
exactly the same way with readline. (In fact, bash substitutes the
ANSI escapes for the bash escapes on output.) The ASCII escapes also
work identically to \[ \] when used directly in a prompt.